backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
pydantic==2.10.5
pydantic-settings==2.7.0
python-dotenv==1.0.1
celery==5.4.0
redis==5.2.1
openai==1.59.8
beautifulsoup4==4.12.3
requests==2.32.3
feedparser==6.0.11
httpx==0.28.1
playwright==1.48.0
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
pgvector==0.2.4
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0